refactor: migrate build system from Webpack to WXT
Replace webpack.config.cjs + hand-authored src/manifest.json with WXT's
convention-based entrypoint discovery and wxt.config.ts:

- Move src/background/background.ts -> src/entrypoints/background.ts
  (wrapped in defineBackground; listener registrations moved into main())
- Move src/popup/* -> src/entrypoints/popup/index.html + popup.ts + styles.css
- Move 9 dynamically-injected content scripts under src/entrypoints/*.ts
  (wrapped in defineUnlistedScript; output paths preserved so existing
  chrome.scripting.executeScript({ files: [...] }) calls still resolve)
- Move all injected CSS into public/ for verbatim copy at the same paths
- Delete webpack.config.cjs and src/manifest.json
- Replace dist/ build target with .output/chrome-mv3/
- Update tsconfig.json to extend ./.wxt/tsconfig.json
- Update .releaserc.json to build via 'bun run zip' and copy the artifact
- Update README with WXT-flavoured development instructions

Side-effect cleanups surfaced by the code review:
- Extract show_detection_banner() helper, consume from detect_domain and
  detect_rss_feed (kills ~22 lines of duplicate banner code per file)
- Use set_setting() in detect_domain / detect_rss_feed instead of
  inlining chrome.storage.sync.get -> mutate -> set
- Split set_setting into its own file so content-script entrypoints don't
  pull in pull_feed / fast-xml-parser (saves ~63 kB per detect_* bundle)
- Collapse mirrored added/removed branches in detect_schooltape into a
  single handle_stylesheet_change() helper
- Trim verbose w_a_r list to glob patterns
- Drop redundant 'compile' npm script
