+++ .releaserc.json
@@ -13,13 +13,13 @@
         [
             "@semantic-release/exec",
             {
-                "prepareCmd": "sed -i 's/\"version\": \".*\"/\"version\": \"${nextRelease.version}\"/' src/manifest.json && bun run build && zip -r extension-chrome.zip dist/"
+                "prepareCmd": "bun run zip && cp .output/*chrome*.zip extension-chrome.zip"
             }
         ],
         [
             "@semantic-release/git",
             {
-                "assets": ["package.json", "CHANGELOG.md", "src/manifest.json"],
+                "assets": ["package.json", "CHANGELOG.md"],
                 "message": "chore(release): ${nextRelease.version} [skip ci]\n\n${nextRelease.notes}"
             }
         ],
