+++ package.json
@@ -2,13 +2,15 @@
   "name": "schoolbox-chrome-extension",
   "version": "1.3.0",
   "description": "A Chrome extension for SchoolBox built with TypeScript, React, and Bun.",
-  "main": "src/background/background.ts",
   "scripts": {
-    "typecheck": "node node_modules/typescript/bin/tsc --noEmit",
-    "build": "webpack --config webpack.config.cjs --mode production",
-    "dev": "webpack --config webpack.config.cjs --watch",
-    "serve-dist": "bun serve dist",
-    "start": "bun run build && bun run serve-dist"
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "zip:firefox": "wxt zip -b firefox",
+    "postinstall": "wxt prepare",
+    "typecheck": "tsc --noEmit"
   },
   "keywords": [
     "chrome",
@@ -20,37 +22,26 @@
   "license": "MIT",
   "dependencies": {
     "@ali-tas/htmldiff-js": "^2.0.1",
-    "@babel/core": "^7.27.1",
-    "@babel/preset-env": "^7.27.2",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
-    "@types/chrome": "^0.1.38",
     "@types/diff": "^8.0.0",
     "@types/html-to-text": "^9.0.4",
-    "babel-loader": "^10.0.0",
-    "copy-webpack-plugin": "^13.0.0",
-    "css-loader": "^7.1.2",
     "diff": "^9.0.0",
     "fast-xml-parser": "^5.2.3",
     "fuse.js": "^7.1.0",
     "html-to-text": "^9.0.5",
-    "html-webpack-plugin": "^5.6.3",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "style-loader": "^4.0.0",
-    "ts-loader": "^9.5.2",
-    "webpack": "^5.99.8",
-    "webpack-cli": "^6.0.1"
+    "react-dom": "^19.1.0"
   },
   "devDependencies": {
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/exec": "^7.1.0",
     "@semantic-release/git": "^10.0.1",
+    "@types/chrome": "^0.1.42",
     "@types/react": "^19.1.3",
     "@types/react-dom": "^19.1.3",
     "bun-types": "^0.1.0",
     "prettier": "3.5.3",
     "semantic-release": "^25.0.3",
-    "typescript": "^6.0.2"
+    "typescript": "^6.0.2",
+    "wxt": "^0.20.0"
   }
 }
