--- src/content/modules/post_history_tracking/content.ts
@@ -1,9 +1,0 @@
-import {setup} from "./setup"
-
-async function main() {
-    await setup()
-}
-
-if (document.querySelector(".meta") && !document.querySelector("#ultrabox-history-container")) {
-    main()
-}
