+++ src/entrypoints/display_history.ts
@@ -1,0 +1,14 @@
+import { setup } from "../content/modules/post_history_tracking/setup"
+
+export default defineUnlistedScript(() => {
+    async function main() {
+        await setup()
+    }
+
+    if (
+        document.querySelector(".meta") &&
+        !document.querySelector("#ultrabox-history-container")
+    ) {
+        main()
+    }
+})
