+++ src/entrypoints/inject_css_tools.ts
@@ -1,0 +1,24 @@
+import { darken } from "../content/modules/launcher/functions/darken"
+
+export default defineUnlistedScript(() => {
+    function set_timetable_bg() {
+        const timetable_elements = Array.from(
+            document.querySelectorAll(".timetable-subject")
+        ) as HTMLDivElement[]
+
+        // check if the timetable has already been darkened
+        if (
+            !timetable_elements.some(
+                el => getComputedStyle(el).backgroundColor === "rgb(204, 255, 204)"
+            )
+        ) {
+            return
+        }
+
+        for (let elem of timetable_elements) {
+            elem.style.backgroundColor = darken(getComputedStyle(elem).backgroundColor, 0.7)
+        }
+    }
+
+    set_timetable_bg()
+})
