+++ src/entrypoints/popup/index.html
@@ -4,7 +4,7 @@
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>UltraBox Extension</title>
-        <link rel="stylesheet" href="styles.css" />
+        <link rel="stylesheet" href="./styles.css" />
     </head>
     <body>
         <div class="container">
@@ -251,6 +251,6 @@
                 </div>
             </div>
         </div>
+        <script type="module" src="./popup.ts"></script>
     </body>
-    <script defer src="popup.js"></script>
 </html>
