+++ src/entrypoints/popup/styles.css
