--- src/manifest.json
@@ -1,24 +1,0 @@
-{
-    "manifest_version": 3,
-    "name": "UltraBox",
-    "version": "1.3.0",
-    "description": "A Chrome extension for SchoolBox.",
-    "permissions": ["storage", "activeTab", "scripting", "tabs", "unlimitedStorage"],
-    "host_permissions": ["<all_urls>"],
-    "background": {
-        "service_worker": "background.js"
-    },
-    "action": {
-        "default_popup": "popup.html",
-        "default_title": "UltraBox"
-    },
-    "commands": {
-        "toggle-launcher": {
-            "suggested_key": {
-                "default": "Ctrl+K",
-                "mac": "Command+K"
-            },
-            "description": "Toggle UltraBox Launcher"
-        }
-    }
-}
